ValveControl basic implementation.
Valves are displayed as a set of buttons, each controller on its own line.
Valves can be highlighted (processing of mouse enter/left events) and opened
or closed (by clicking on them). Valve state is periodically queried from
VcDriver. ValveControl panel automatically updates itself if it detects that
the number of valves changed.